feat: create separate Component for data product id
